requirements.txt
Adafruit-GPIO >= 1.0.3
Adafruit-PureIO >= 0.2.3
apa102-pi >= 2.2.0
Click >= 7.0
edgetpu >= 1.2.0
Flask >= 1.1.1
Flask-SocketIO >= 4.2.1
itsdangerous >= 1.1.0
Jinja2 >= 2.10.1
MarkupSafe >= 1.1.1
numpy >= 1.12.1
picamera >= 1.13
pygame >= 2.0.0.dev4
python-engineio >= 3.9.3
python-socketio >= 4.3.1
RPi.GPIO >= 0.7.0
spidev >= 3.4
Werkzeug >= 0.16.0
